+++ pages/community/index.js
@@ -24,7 +24,7 @@ import Layout from '../../layouts/LayoutCommunity';
 import Newsletter from '../../components/community/newsletter/Newsletter';
 import Events from '../../components/community/events/Events';
 import Resources from '../../components/community/resources/Resources';
-import Slack from '../../components/community/slack/Slack';
+import Discord from '../../components/community/discord/Discord';
 import TechTalk from '../../components/community/tech-talk/TechTalk';
 import GetInvolved from '../../components/community/get-involved/GetInvolved';
 import Contact from '../../components/community/contact/Contact';
@@ -94,7 +94,7 @@ export default function Community() {
           </Row>
 
           <Row className="pageContentRow communityRow slackRow">
-            <Slack />
+            <Discord />
           </Row>
 
           <Row className="pageContentRow communityRow">
